Add searching with context text to client
This change enhances the output of the archive client so that it is
easier to both lookup the issue/page and determine whether a
particular result is relevant. Result relevance is displayed by adding
a little context around the search query within the result text, not
unlike the bolded text in a Google search result.