Refactor and re-do how the CLI works
This change modifies how the CLI is organized in anticipation of
future functionality. Instead of being single function, the command
line interface is based off of an ACTION NOUN (eg 'search' and 'text')
interface using multiple sub-parsers. This allows additional
functionality to be added without too much fuss while making it easy
to understand how to use the script.